// src/calculo_diarias/diarias-enum.ts
export enum Acompanha {
  DENTRO = 766.22,
  FORA = 851.36,
  INTERNACIONAL = 472.55,
}

export enum Local {
  DENTRO = 'dentro',
  FORA = 'fora',
  INTERNACIONAL = 'internacional',
}

export enum Municipios {
  MACAPA = 'Macapá',
  MAZAGAO = 'Mazagão',
  SANTANA = 'Santana',
}

export enum UF {
  AC = 'AC',
  AL = 'AL',
  AM = 'AM',
  AP = 'AP',
  BA = 'BA',
  CE = 'CE',
  DF = 'DF',
  ES = 'ES',
  GO = 'GO',
  MA = 'MA',
  MG = 'MG',
  MS = 'MS',
  MT = 'MT',
  PA = 'PA',
  PB = 'PB',
  PE = 'PE',
  PI = 'PI',
  PR = 'PR',
  RN = 'RN',
  RJ = 'RJ',
  RO = 'RO',
  RR = 'RR',
  RS = 'RS',
  SC = 'SC',
  SE = 'SE',
  SP = 'SP',
  TO = 'TO',
}